rearrange the layout of Assets/Puerts to match the package layout
