+++ unity/.gitignore
@@ -1,4 +1,4 @@
-*.meta
+# *.meta
 /general/Bin
 /general/Lib
 /general/vs2013/*/
