+++ unity/Assets/Puerts/Runtime/Src/StaticTranslate.cs
